refactor methods from python client to abstract python codegen
